fix(workflows): break host-local-workflows import cycle that crashed binary dispatch
`host-local-workflows.ts` statically imported helpers from `auto-dispatch.ts`,
whose module-bottom TLA dynamic-imports `runtime/orchestrator-entry.ts` —
which in turn imports `lookupLocalWorkflow` back from `host-local-workflows.ts`
while it's still suspended on its own first import. In compiled binaries the
SDK barrel is on the import path, so this cycle fired on every
`_orchestrator-entry` dispatch: `localWorkflowRegistry` was in TDZ when
`lookupLocalWorkflow` ran, the orchestrator pane crashed with
`TypeError: undefined is not an object`, and the parent CLI silently exited 0.

Dev mode dispatched through `packages/atomic-sdk/src/cli.ts` (which only
imports `auto-dispatch.ts`), so `host-local-workflows.ts` was never
mid-evaluation — hence the dev-vs-binary divergence introduced in 7a55ab41.

Fix: extract `validateDispatchToken`, `findSub`, `parseAtomicRunArgv`, and
`AtomicRunArgs` into a new `lib/dispatch-utils.ts` (no TLA, no SDK imports).
`host-local-workflows.ts` now imports from there; `auto-dispatch.ts`
re-exports for any external consumer that reached for them via that path.

CI hardening so the next dev-vs-binary regression doesn't slip through:
- `ci.yml` `checks` job now builds the host binary before `bun test`, so
  binary-gated suites actually run instead of silently skipping.
- `locateBuiltBinary()` in `orchestrator-entry.test.ts` throws under
  `process.env.CI` when no binary is present — a missing build is a hard
  failure, not a silent green.
- Both existing tests now also assert
  `not.toContain("undefined is not an object")` and
  `not.toMatch(/TypeError|ReferenceError/)`, so a future module-evaluation
  regression on the dispatch path fails with a clear signal.
