+++ packages/atomic-sdk/src/runtime/orchestrator-entry.test.ts
@@ -1,27 +1,33 @@
 /**
  * Integration test: `_orchestrator-entry` against the real compiled binary.
  *
- * Regression guard for the dev-vs-binary divergence where Bun's
- * `bun build --compile` collapses every bundled module's
- * `import.meta.path` to the binary's bunfs entry path
- * (`/$bunfs/root/<binary>`), so the `definition.source` captured at
- * workflow-module-eval time pointed at the binary itself. Before the
- * fix, dynamic-importing that path re-loaded `cli.ts` (no default
- * export) and the orchestrator pane died with
- * `"does not export a valid WorkflowDefinition"` while the launcher
- * shell exited 0 — invisible to the chat-style smoke tests.
+ * Catches dev-vs-binary divergences in the orchestrator dispatch path.
+ * Two known prior regressions this guards against:
  *
- * The fix routes binary-mode invocations through the builtin registry
- * by `name + agent`, falling back to dynamic-import for dev /
- * installed-package mode. This test exec's the real binary with the
- * post-fix launcher contract and asserts:
- *   1. The bug signature ("does not export a valid WorkflowDefinition")
- *      does not appear.
- *   2. The action gets past workflow resolution (it complains about
- *      missing ATOMIC_WF_* env, which is the next failure mode in
- *      `runOrchestrator`).
- *   3. An unknown workflow name surfaces a clean registry-miss error,
- *      not the old import-failure error.
+ *   1. `import.meta.path` collapse (bunfs): `bun build --compile`
+ *      points every bundled module's `import.meta.path` at the binary
+ *      entry, so the captured `definition.source` re-imported cli.ts
+ *      and the pane died with "does not export a valid WorkflowDefinition"
+ *      while the launcher exited 0 (silent failure).
+ *
+ *   2. Static import cycle in the SDK barrel: `host-local-workflows.ts`
+ *      → `auto-dispatch.ts` (TLA) → `orchestrator-entry.ts` →
+ *      `host-local-workflows.ts` (still suspended). The orchestrator
+ *      called `lookupLocalWorkflow` while `localWorkflowRegistry` was in
+ *      TDZ → "TypeError: undefined is not an object". Same silent-exit
+ *      shape (parent CLI returns 0). Fixed by extracting helpers into
+ *      `lib/dispatch-utils.ts` so the cycle is broken.
+ *
+ * Both bugs only manifest in compiled binaries because dev mode
+ * dispatches through the SDK's standalone cli.ts (no barrel re-export),
+ * and both surface as the orchestrator pane crashing while the parent
+ * exits 0. The assertions below combine the two bug signatures: any new
+ * "TypeError"/"ReferenceError" / "is not an object" / "does not export
+ * a valid WorkflowDefinition" output means a regression.
+ *
+ * `locateBuiltBinary()` throws (rather than returning null) when run
+ * under CI without a built binary — otherwise these tests would silently
+ * skip on every PR and the regression class would slip through.
  */
 import { test, expect, describe } from "bun:test";
 import { existsSync } from "node:fs";
@@ -48,10 +54,17 @@ describe.skipIf(!builtBinary)("compiled binary _orchestrator-entry", () => {
 
     const out = result.stdout.toString() + result.stderr.toString();
 
-    // Bug signature: pre-fix, the dynamic import re-loaded cli.ts and
-    // threw InvalidWorkflowError. The new path never imports the
-    // source, so this string must not appear.
+    // Bug signatures from the two known prior regressions (see file header).
+    // Pre-fix #1: the dynamic import re-loaded cli.ts and threw
+    // InvalidWorkflowError → "does not export a valid WorkflowDefinition".
     expect(out).not.toContain("does not export a valid WorkflowDefinition");
+    // Pre-fix #2: TDZ in localWorkflowRegistry due to static import cycle
+    // through the SDK barrel.
+    expect(out).not.toContain("undefined is not an object");
+    // Catch-all for any new module-evaluation crash on the dispatch path —
+    // a clean miss should surface a domain error (registry/env), not a
+    // runtime exception class.
+    expect(out).not.toMatch(/TypeError|ReferenceError/);
 
     // Positive signal: we got past workflow resolution into
     // runOrchestrator → validateOrchestratorEnv. The next failure mode
@@ -73,6 +86,11 @@ describe.skipIf(!builtBinary)("compiled binary _orchestrator-entry", () => {
     expect(out).toContain("no-such-workflow");
     expect(out).toContain("builtin registry");
     expect(result.exitCode).not.toBe(0);
+    // Same module-evaluation regression guards as the registry-hit case —
+    // a registry miss must come back as a clean domain error, not a
+    // runtime exception thrown from a broken import graph.
+    expect(out).not.toContain("undefined is not an object");
+    expect(out).not.toMatch(/TypeError|ReferenceError/);
   });
 });
 
@@ -92,6 +110,17 @@ function stripWorkflowEnv(env: NodeJS.ProcessEnv): Record<string, string> {
   return out;
 }
 
+/**
+ * Resolve the host's built atomic binary, or return `null` so the
+ * `describe.skipIf(!builtBinary)` gate skips the suite locally.
+ *
+ * In CI we throw instead of returning `null`. These tests are the only
+ * regression guard for an entire class of dev-vs-binary divergence
+ * (silent orchestrator-pane crashes while the parent exits 0); a silent
+ * skip on a CI runner that forgot the build step would re-open that
+ * gap. The `checks` job in `.github/workflows/ci.yml` is responsible
+ * for running `bun packages/atomic/script/build.ts` before `bun test`.
+ */
 function locateBuiltBinary(): string | null {
   const ext = process.platform === "win32" ? ".exe" : "";
   let dir = import.meta.dir;
@@ -109,9 +138,23 @@ function locateBuiltBinary(): string | null {
         const candidate = join(distRoot, target, "bin", `atomic${ext}`);
         if (existsSync(candidate)) return candidate;
       }
+      if (process.env.CI) {
+        throw new Error(
+          `[orchestrator-entry.test] No built atomic binary found under ${distRoot}. ` +
+            `Run \`bun packages/atomic/script/build.ts\` before \`bun test\` in CI — ` +
+            `silently skipping these tests on a CI runner re-opens the dev-vs-binary ` +
+            `regression gap they exist to catch.`,
+        );
+      }
       return null;
     }
     dir = dirname(dir);
+  }
+  if (process.env.CI) {
+    throw new Error(
+      `[orchestrator-entry.test] Could not locate workspace root from ${import.meta.dir}; ` +
+        `binary-gated tests cannot run in CI without a build.`,
+    );
   }
   return null;
 }
